// src/app/(frontend)/recipes/[slug]/page.tsx
import type { Metadata } from 'next'

import { RelatedRecipes } from '@/blocks/RelatedRecipes/Component'
import { PayloadRedirects } from '@/components/PayloadRedirects'
import configPromise from '@payload-config'
import { getPayloadHMR } from '@payloadcms/next/utilities'
import { draftMode } from 'next/headers'
import React, { cache } from 'react'
import RichText from '@/components/RichText'

import type { Recipe } from '@/payload-types'

import { RecipeHero } from '@/heros/RecipeHero'
import { generateMeta } from '@/utilities/generateMeta'
import PageClient from './page.client'

export async function generateStaticParams() {
  const payload = await getPayloadHMR({ config: configPromise })
  const recipes = await payload.find({
    collection: 'recipes',
    draft: false,
    limit: 1000,
    overrideAccess: false,
  })

  const params = recipes.docs.map(({ slug }) => {
    return { slug }
  })

  return params
}

type Args = {
  params: Promise<{
    slug?: string
  }>
}

export default async function Recipe({ params: paramsPromise }: Args) {
  const { slug = '' } = await paramsPromise
  const url = '/recipes/' + slug
  const recipe = await queryRecipeBySlug({ slug })

  if (!recipe) return <PayloadRedirects url={url} />

  return (
    <article className="pb-16">
      <PageClient />

      {/* Allows redirects for valid pages too */}
      <PayloadRedirects disableNotFound url={url} />

      <RecipeHero recipe={recipe} />

      <div className="flex flex-col items-center gap-4 pt-8">
        <div className="container lg:mx-0 lg:grid lg:grid-cols-[1fr_48rem_1fr] grid-rows-[1fr]">
          <RichText
            className="lg:grid lg:grid-cols-subgrid col-start-1 col-span-3 grid-rows-[1fr]"
            content={recipe.content}
            enableGutter={false}
          />
        </div>

        {recipe.relatedRecipes && recipe.relatedRecipes.length > 0 && (
          <RelatedRecipes
            className="mt-12"
            docs={recipe.relatedRecipes.filter((recipe) => typeof recipe === 'object')}
          />
        )}
      </div>
    </article>
  )
}

export async function generateMetadata({ params: paramsPromise }: Args): Promise<Metadata> {
  const { slug = '' } = await paramsPromise
  const recipe = await queryRecipeBySlug({ slug })

  return generateMeta({ doc: recipe })
}

const queryRecipeBySlug = cache(async ({ slug }: { slug: string }) => {
  const { isEnabled: draft } = await draftMode()

  const payload = await getPayloadHMR({ config: configPromise })

  const result = await payload.find({
    collection: 'recipes',
    draft,
    limit: 1,
    overrideAccess: draft,
    where: {
      slug: {
        equals: slug,
      },
    },
  })

  return result.docs?.[0] || null
})
